Fix data corruption in read prefetch logic (#43)
When a read request was mostly, but not completely, satisfied out of
prefetch buffer, data corruption could occur. This depended on the
specific timing of requests and was quite rare.

Root cause was that the next prefetch started before checking whether
all prefetch data for current request had been written out.