RP2040: Fix crash when responding to WDTR request.
The RP2040 accelerated SCSI routine in refill_dmabuf() accesses data in 16-bit words for speed.
Cortex-M0 crashes on unaligned access.

Most SCSI data transfer requests are from aligned RAM buffer,
but the responses to some negotiation messages are constants from flash.
In this case the negative response to Wide Data Transfer Request was
unaligned in flash.

Added a check for buffer alignment and use the slower byte-per-byte transfer
for unaligned buffers. Because it is just a few bytes speed doesn't matter here.