SCSI-1: Add delay before changing to DATA phase for compatibility with Akai S1000 (#71)
It seems that changing SCSI control signals too quickly after command is received
confuses some SCSI hosts. This commit adds a 400 us delay which has been determined
to be enough for Akai S1000, where the problem was observed.

To avoid performance degradation, the extra delay is performed in parallel with
SD card access. The SD card typically has a 1 millisecond latency, so the extra
wait does not increase the total latency.

The delay is only active in SCSI-1 mode. This is activated either explicitly
by setting EnableSCSI2 = 0, or automatically if the host does not send the
SCSI-2 Identify message.