GD32: Handle timeouts when waiting for SDIO response (#82)
Previously the GD32 SDIO driver for some reason only had a timeout
path for R7 response type. Copied the same pattern to response types
R1-R6.

Still unsure what is causing the timeouts in a way that is not detected
by the hardware timeout mechanism. But this commit should at least allow
to recover from the situation automatically.