Limit drive capacity to actual file size (#80)
In versions 1.0.8 and 1.0.9, the full allocated size of the file
was used instead of its actual size. Because FAT allocates space
typically in 32 kB blocks, this could be larger than the actual
size of the image.

This fix can affect drives formatted using firmware versions 1.0.8
and 1.0.9 if the image file size was different than the allocated
size. A notice is printed in that case and the old behavior can
be restored with config setting [SCSI] UseFATAllocSize = 1