RP2040: Fix hang with ModeSense command with synchronous mode on Amiga
The RP2040 accelerated code works in multiples of two bytes,
so a slower fallback code path is used to transfer command
responses that have an odd number of bytes.

This fallback code wasn't fully reliable with synchronous mode because
if an interrupt occurred in the middle, the response ACK pulse could be
missed.

Fixed by using falling edge interrupt register to detect ACK pulses.