Rewrite GPIO pin inits
Most of the work was spent rewriting GPIO pin inits and getting
rid of Greenpak code. Some work was done on the sdio sdcard code
but more work needs to be done before the project compiles.

It was found that updating the flash code will not work because
page erases are not support with the current chip. Switching to
sector erasures will most likely fix the issue.