Firmware compiles and links
Was able to get the firmware to link. There are a few outstanding
issues to take care of. They can be found by doing a project wide search
for '@TODO'.

Mainly erasing the flash to load firmware off the SD card.

There is also a DMA memory to memory transfer that needs to be
figured out.

And how to properly use GPIO_CTL for the F4xx compared to F20x's
GPIO_CTL0