RP2040: Make accelerated SCSI routines work for odd number of bytes
Primarily this fixes issues with synchronous mode transfers of
ModeSense data. Previous fix in #89 (commit f11c0d4750) was a
workaround that didn't fully implement the sync mode behavior.

This commit unifies the behavior so that all writes and read go
through the same code path.

As-is, this commit causes a small performance degradation for
transfer rates above 7 MB/s.
It will be optimized in a following commit.