Add the ability to disable the status LED
The LED is only disabled after the SD card is read.

That means it will blink on boot and blink until it finds
an SD card that is readable with a the zuluscsi.ini file and the
"DisableStatusLED = 1" under "[SCSI]" is set. After which it will
never blink again until the board is power cycled without
"DisableStatusLED = 1" flag.