Per the GPL's FAQ: https://www.gnu.org/licenses/gpl-faq.en.html#NoticeInSourceFile
I have added the appropreate GPL header to each file. If there was an
existing copyright line it was kept.

Since there are so many contributors through these projects I've elected
to put them all in a CONTRIBUTORS.txt file. This file will be updated as
new contributors add to this project.

There is no intent to remove any copyright notices. Care was taken
to ensure each file in this repository is correctly licensed. If any
mistakes were made in this or future commits I will address them as soon
as possible. Any changes or clarifications by upstream projects will be
incorperated.

Files autogenerated by tools or derived from the Raspery Pi foundation
were left as is. It is unclear how the licensing should work in
these cases.