Open image as read-only if attribute set
A disk image will be opened as read-only if the read-only attribute
is set on the image file. If the SCSI disk is been written to a
read-only image, the "Write protected" sense code will be returned.
If that is somehow bypassed the ImageBackingStore write method will
log an error and return 0 bytes written.

Issue #124