Save debug log even when no SCSI requests are coming in
Normally we wait for the STATUS phase to be able to save
log file without disturbing SCSI processing. But this can
cause the very last log messages not be saved to file.

This commit changes behavior only when debug log is enabled.
Log will be saved after 2 seconds when no new SCSI requests are coming.
Shouldn't affect functional behavior, but to limit any performance impacts it is
limited to debug mode.