Refactor flash programmer
Removed as much custom code from the ZuluSCSI_bootloader.cpp for the
F4 into ZuluSCSI_platform.cpp. In ZuluSCSI_platform.cpp, created
a generic sector eraser (erase_flash_sector) and a generic flash
writer (write_flash) that can be reused for something else.

Because there are now two ways to erase the flash in the
ZuluSCSI_bootloader.cpp (page and sector) for all the environments,
it might be worth refactoring out that specific detail for each
environment.