Prevent START STOP UNIT from changing SCSI2SD `started` value.
This fixes an issue where Macs on reboot would stop the drive and never try to restart it. This also matches observed behavior on a Matshita drive, which would just spin down in response to this command and continue to indicate a good response to TEST UNIT READY as long as a disk was in the drive.