Fix PhyMode=1(PIO) on ZuluSCSI GD32 V1.1 (fix #245)
With the GD32 SDIO driver, check_sd_read_done() was returning true
even when no SD card read was in progress. This is because unlike
the SPI driver, the SDIO driver does not disable the DMA channel.

Also fixed a bug where scsiIsWriteFinished() erroneously returned
true on this early return, causing this to be a data corruption
instead of a hang.

Only affects GD32 V1.1 platform when PhyMode=1 setting is specified
in zuluscsi.ini.