Temporary fix for DMA_TIMER mode on ZuluSCSI v1.1
The ZuluSCSI v1.1 in DMA_TIMER mode would only boot to the Welcome
screen before freezing on a Mac LC III. Setting the prefetch
buffer to 4k seems to fix the issue, but doesn't address the root cause.

This fix sets the limit of the prefetch buffer to 4k only in,
PhyMode = 2, DMA_TIMER mode, so the other modes can take advantage of
of the full size of the prefetch buffer.