Add Mac sanity checks
The main purpose of this code is to emit warnings when Mac quirks are
turned on and the image(s) are disks.

The warnings are emitted if any of the following checks fail:
    - The disk image contain Apple Magic signature
    - The HFS block size is correct
    - The Mac SCSI driver is smaller than or equal to the max size
    - The MAC SCSI driver is smaller than or equal to the image size
    - The Lido driver is not installed

The original implementation put the sanity checks under platform code.
Since none of the code is platform specific, it has been moved
to the root src directory so all build targets can take advantage
of the sanity checks.

Initial code authored by Eric Helgeson and taken from:
https://github.com/BlueSCSI/BlueSCSI-v2/commit/cc1ff5f4d03ce976d01ead09cee1106ce0acae08
https://github.com/BlueSCSI/BlueSCSI-v2/commit/d00ebe016b8c7ef4777c31cef9cd002043de7d46

Co-authored-by:  Eric Helgeson <erichelgeson@gmail.com>