Works on boot and the LED works also
The scsi_accel_target.cpp was clearing out a PIO instruction memory
that was causing the Wireless SPI pio state machine code to crash.

Removing that line conditionally fixed the issue. It can also
boot with both a disk image and a DaynaPORT device, but for some reason
the WiFi doesn't connect. So in its current state it is a network
device only.