Add serial logging to GD32F205 builds
Serial logging should work for all PlatformIO environments
based on the GD32F205.

There is an unresolved issue where the the GD32F205 platform library
is being pulled into the RP2040 builds. The quick fix is
to add a lib_ignore to the GD32F205 platform library for RP2040 builds.
This should be figured out though.