Set network frames to 25 and SCSI2SD buffer to 4x
Was able to set the network frames to 24 and reduce the SCSI2SD buffer
from 8x to 4x. This gave enough SRAM to reenable USB serial logging
and add back the prefetch buffer to 2048 bytes.

Sometimes the download stops when creating a new file with data
already in Netscape 2.0 download buffer. It seem to happen
only on the first download, but more testing around this issue
is needed.