Add 3 modes image writing in initiator mode
In the zuluscsi.ini file a new option `InitiatorImageHandling`
has been added.
The values are:
 - 0: Stop if image exist, an image file exist don't re-image the drive
 - 1: Create new image, an image file exists re-image to a new file
 - 2: Overwrite image, an image file exists overwrite that file

The default is 0

Also added the option `InitiatorID` to zuluscsi.ini which sets the
initiator SCSI ID. The default setting for the host (initiator)
SCSI ID to 7, when the option is used the host SCSI ID is set to that
number. If the value is out of range, it sets it to 7.