Skip image if larger than max file size
If the target image is larger than the max filesize of the file system
E.g. 4GB on a FAT32 SD card, instead of truncating the image
to the max filesize, skip imaging and emit an error message.

Also fixed an error where a `=` should have been an `|=`
when testing for not enough free space on the SD card.

Modified version of:
https://github.com/BlueSCSI/BlueSCSI-v2/pull/80/commits/8e9a3c0bfc2c479901b1c7a5772c0e3c9f4d9ca8

Co-authored-by: Eric Helgeson <erichelgeson@gmail.com>