Credit to Joshua Stein (@jcs)
This commit is to recognize the contribution Joshua Stein made to the
BlueSCSI V2 project to create a SCSI DaynaPORT Wi-Fi Ethernet emulator.

@jcs's code was adapted for the ZuluSCSI Pico board for
this project at commit #2b9879f542972eb55ca26cf19f1739d699c6e179
But wasn't tagged as a co-author. The purpose of this commit is to
recognize the significant effort that went in to writing this code.

The original pull request, which @jcs contributed to BlueSCSI V2,
can be viewed at https://github.com/BlueSCSI/BlueSCSI-v2/pull/59

Co-authored-by: joshua stein <jcs@jcs.org>