Fix device type being ignored in zuluscsi.ini
Bug raised in https://github.com/ZuluSCSI/ZuluSCSI-firmware/issues/325
Seems to be that the device type wasn't being passed to the
`scsiDiskOpenHDDImage` function in two cases. One where CD when
switching to new CDs. Two, in the case of the issue raised, when a
SCSI ID  is defined in the in `zuluscsi.ini` file. Because they
were not being passed the device type was defaulting to
fixed disk (hdd).

Haven't analyzed how the issue was introduced, but a guess is that
it was a change to `scsiDiskOpenHDDImage` behavior.