Fix an existing bug for TIMER DMA phy mode
The waiting for the last ACK to go low in `scsi_accel_dma.cpp:388`
timeout was too quick for an Adaptec AHA-2940AU. The timeout was based
on the iteration count through a tight loop. It has been switched
to a 500 millisecond timeout check.

Comments about right aligned vendor and product id SCSI strings
have been removed as they are misleading.