Fix device type for IMG0-9
This is for issue https://github.com/ZuluSCSI/ZuluSCSI-firmware/issues/345
The bug was introduced in https://github.com/ZuluSCSI/ZuluSCSI-firmware/releases/tag/v2023.10.20

This fixes an issue when the device specific options in the
`zuluscsi.ini` file `IMG0`-`IMG9` are set along with
`Type` set to anything besides 0. The type would be ignored
and the device would always be set to a fixed device (Type=0).

The underlying fix was to add a new device type `S2S_CFG_NOT_SET`,
which is only used in initialization to determine where the device
type should be overwritten.