Fix SCSI ID reporting using dip switch setting
There was a bug where the hardware preset settings were being applied
no matter if direct/raw mode was enabled or disabled and the dip switch
set SCSI ID was being reported in the wrong mode.

This fixes the bug and reports the SCSI ID in the correct mode.
It also changes the text of the DIP switch settings.