Fix device selector
The init of the direct mode saved the device selection to the
settings struct directly. This happened before the settings
were initialized, so the device selection was lost.

The code already set the device mode from a member variable
in the class in platform_hw_config.cpp after the setting struct was
initialized. So a simple switch from saving the device mode directly in
the settings struct to storing it in the member variable fixed the
issue.

Also switched the debug text in direct mode mentioning the image file's
start and end sectors on the SD card to referring to the device spanning
the start and end sectors of the SD card.