Attempt at supporting the vendor 0xD8 command
This is a CD data read that is mainly for Plextor CDROM drives.
It is a vendor CD Data read used mainly in duplicating CDs.
It support reading more raw data then the other CD Read commands
for better CD cloning.

This is an attempt to address issue https://github.com/ZuluSCSI/ZuluSCSI-firmware/issues/258