Make aligned reads from the SD card for the D8 cmd
The code  now fetches data from the SD card aligned on 512 byte
boundaries. This means prefetching and post-fetching from the SD card
so the unaligned 2352 byte CD sector size can properly fit in the
buffer. To be written over the SCSI bus.

The end case where the last CD sector is read hasn't been addressed yet,
so in theory the buffering code could attempt to read past the end
of the SD card.