Fix setting type with default SCSI strings
This fixes an issue where a device type is set in a SCSI ID section
in the `zuluscsi.ini` did not apply the default vendor, product,
and revision SCSI strings, e.g.
```
[SCSI3]
Type = 2 # CDROM type
```
This was reported with the standard `HARDDRIVE` product string and
not `CDROM`. The fix now reads the device type from `zuluscsi.ini`
before default types are applied. It moves it out of the
`readIniSCSIDeviceSetting` function unfortunately but is the
easier fix than reapplying the default SCSI strings after `Type` is
read.