Dynamic linker script and fix ROMs for DaynaPORT
Converted the linker script to a Python string template.
It fills in the path value for the object files that are put into flash.
It also fills in the maximum flash allocated to the program.

The motivation for the different flash sizes is because the DaynaPORT
takes up more flash than allocated to the other RP2040 base builds.
To make the RP2040 build bigger to the same size of the DaynaPORT build
will break users the have a ROM drive installed on their ZuluSCSI. So
the code bases are built with different flash sizes.

Also fixed now is the DaynaPORT code now writes the ROM at the
program boundary instead of partially overwriting the code in flash.

The program boundary is set in the `platformio.ini` file with the
`program_flash_allocation` setting.