Adding Mass Storage to the F4 target
Enabled High Speed USB to the F4 target. In alpha stage. There is
something wrong with SDIO interrupt when going into MSC mode.
A static global variable `transend`` doesn't seem to get updated but the
flag for the interrupt is set the sets it is raised.

For now the code is polling the interrupt handler so the flag can be
cleared and the `transend` variable set.

Search for this comment:

`// \todo Figure out why the SDIO interrupt isn't firing`

To find where the interrupt handler is being polled.