Fix starting offset bug
A missing `else if` is causing the wrong offset being read during
a Plextor 0xD8 vendor SCSI CDB. Adding the `else` fixes the issue.
It may or may not address the issue https://github.com/ZuluSCSI/ZuluSCSI-firmware/issues/258

Also dropped the `scsi_id` parameter for the function
`scsiDiskOpenHDDImage` as it is redundant, with `target_idx` parameter
being the same value.