Switch between the same prefix on eject
Add the ability to eject and load filenames with the same prefix.
For example if there are two files "zp2-a.img" and "zp2-b.img"
Them machine will boot with "zp2-a.img" and when ejected will load
"zp2-b.img". It will loop back to "zp2-a.img" when ejected for a second
time.

Ejecting zip drives works both in Linux and Mac OS 7. With debug on
a v1.2 there was a timeout on a write. Did not occur with debug off.