Add reporting for eject button
There is basic messaging around using an eject button for removable
drives.  On board with I2C can be hijacked for Eject GPIOs or with
the v1.2 the is an eject button. This will list the devices
that the eject button(s) are hooked into or messaging about where
to look up instructions for setting up eject buttons.