Do not open and write to a log file in raw mode
This fix stops the opening of both the `zululog.txt` and `zuluerr.txt`
when the drive is in raw mode.

It also renames a variable and function israw to iscontiguous as it
aligns with it actual meaning and adds israw to denote SD card is
accessed via raw mode

It also fixes an unrelated issue where a for loop variable wasn't being
initialized.