Move to latest arduino pico framework and pico-sdk
Moved the RP2040 and RP2350 to the same customized arduino pico
framework. That is compatible with both USB mass storage and the
DaynaPORT Pico W SPI pio Wifi interface.

Other changes include a base PlatformIO target that all RP builds
inherit from. This gets rid of a lot of repeated settings.

An unrelated change was the added the ability to set the timeout for USB
enumeration as sometimes RP devices take longer than second to enumerate
when on a USB hub. And adding both `EnableUSBMassStorage` and the new
`USBMassStorageWaitPeriod` to the example `zuluscsi.ini` file.