Reclocking works across the ZuluSCSI RP line
In `lib/ZuluSCSI_platform_RP2MCU/timings.c` are the three clock settings
for the product line. With a special setting for the SD clock rate of
the Pico 2 as it could not do 50MHz SDIO clock at a 250MHz system clock.
Was able to consolidate all the sdio pio and scsi_accel_target pio
assembly files into one file for each function by setting the gpio
waits dynamically via rewriting the the command in code.

There are 3 things left to do.
 - Consolidate the scsi_accel_host pio files
 - Turn the Audio Build back on
 - See about DaynaPORT compatibility at high clock rates