Use single zip file to update any ZuluSCSI
This will update the ZuluSCSI by placing a zip file with the prefix
`zuluscsi-fw` in the root directory of the SD card.

The firmware will extract the appropriate bin file from the container
zip, delete the container zip and then reboot the board. Once the
board has been rebooted, the bootloader installs the bin file.

This is a initial attempt. Currently if the zip file is compressed
with deflate, the code will install the compressed data which will
crash the board.