Fix ZuluSCSI v1.1plus firmware from not starting
The ZuluSCSI v1.1plus ran out of free SRAM and was overwriting some
variables. The solution was to halve the logging buffer to 8K from 16k.

Also in debug mode the exact error number for the no SD card error was
spamming the log. Now it only shows the error every five seconds when it
is retrying or if the error number changes.