Fix ordering of speed grades
The order of the speed grade text and enum didn't match so speed grade,
so e.g. TurboA wasn't matching the enum SPEED_GRADE_A, etc.

Also set the 1MHz div for SDIO from 50 to 30 for the 250MHz system
clock, as this seems to make some SD cards work with the Pico 2.