Basic tape file marking with Read support
If a named directory like "tp1 - with file marks" is in the root
directory on the SD card. It will set a tape drive on SCSI id 1.
It will look at valid files in the directory and treat them as a single
tape with file markers in between in alphabetical order.
For example
```
tp3_multi-marked-tape
   |
   --- segment.001
   |
   --- segment.002
   |
   --- z_end_segment
```

Will be treated as a tape
```
[Begin | segment.001 data | filemark 0 | segment.002 data | filemark 1 | z_end_segment | filemark 2 | End]
```

Tested under Linux using `dd`, remember to use device `/dev/ntpX`
the prefix 'n' tells linux not to rewind after the device has been read.
'X' is the device ID, probably 0 unless there are more than 1 tape
drive. After the first dd, the tape will be moved to the next segment.
and the next dd will read that segment. `mt rewind` will set the
tape back to the first segment.

Currently reading past the last segment with dd will simply read a
0 byte file. Not sure how to handle this yet.