ImageBackingStore: contiguous optimizations were sometimes skipped
There was an off-by-one error where contiguous image file optimizations
were not applied if the allocated FAT/exFAT extent was exactly the same
size as the image file.

The log message used different logic so it would still report the image
as contiguous (which it is), even though the optimizations were not active.

Actual performance impact is quite small.