DaynaPORT SCSI/Link support for ZuluSCSI Blaster
Support is enabled via optional Raspberry Pi RM2 based add on module
that mimics the Pico 1W/2W's onboard wireless support.

Switched the variant in the board `zuluscsi_blaster.json` file to be
rpipico2w so the proper static libraries with wifi support get linked.

RM2 works with an audio compatible clock speed of around 155MHz.
So the Blaster now initializes around 155MHz then if it doesn't
detect an RM2 up clocks to 200.5MHz.

Turn on the RM2 module's LED and the Pico 1W/2W LEDs (not the activity
LED) early on in the init.

This indicates a couple of things to the user
 - The devices are setup in network mode
 - The communication interface between the ZuluSCSI and the wireless
   module works

The RM2 and Pico W LED turns off when either the network connection
is attempted, in which case the LED goes through some blinking sequences
or there is no network device configured on a SCSI ID.