RP2MCU: Fix USB console not connecting
There was a regression caused by commit 7238526d551f0f3 which
caused the USB console not to connect after the first SCSI request
had been processed.

This seems to be related to the USB request timeout of 500 ms, where
the polling interval was previously 1000 ms.

I didn't research deep enough to find why this works before the first
SCSI request, nor why it doesn't work afterwards. The USB IRQ still
seems to be called so it has something to do with the userspace
processing of requests. Possibly the USB MSC support also affects
it.

In any case, lowering the polling interval to 50 ms fixed the issue
and gives just as good performance.