Use init delay of MSC to fix USB serial connection
It was found that a simple delay was needed before configuring the
MSC TinyUSB callbacks. This allows an immediate USB serial CDC
connection and decreases the initial time to start responding to
USB MSC host commands.

Added a `zuluscsi.ini` setting `InitiatorMSCInitDelay` so the
delay can be customized. Not ideal because the standard way to get into
initiator MSC mode is to have no SD card in the slot, meaning no
`zuluscsi.ini` to read the setting from.