Wrap Read TOC special handling into Mac quirks
This is to fix an issue reported https://github.com/ZuluSCSI/ZuluSCSI-firmware/issues/583
Special handling of the control byte in the Read TOC SCSI CDB was
causing CD-ROM emulation to fail in DOS with a Tekram DC 390 SCSI
controller.

Since the special handling was related to Apple drivers in this commit
https://github.com/ZuluSCSI/ZuluSCSI-firmware/commit/ed3ebe0bbf9c4fbcf71c3861c04846431cde83f3
it was decided to only do the special handling when Mac quirks are
enabled.