Speedup tape reads and fix a reported issue
This speeds up the tape reads to near disk read speeds instead
of around half. On linux it will do two zero reads at the end of
the tape and on the third Linux post an error.

Which is normal behavior according to this line
"In both cases end of data is signified by returning zero bytes for
two consecutive reads." in the Linux kernel doc for SCSI tape:
https://docs.kernel.org/scsi/st.html

It also fixes a bug caught in code review where the sense data
for the tape drive was logically wrong.