Fix Read Defect Data response
Information found from SCSI 2 Spec 9.2.8
The second byte of the Read Defect Data defect list should come from
the third byte of the Read Defect Data command. The defect list's second
byte, which is Reserved, Plist, Glist, Defect List Format was coming
from the second byte of the command, which is LUN, Reserved, and is
incorrect.

This is an attempt to address issue: https://github.com/ZuluSCSI/ZuluSCSI-firmware/issues/634