CLEAR_ROM file was crashing
Placing a CLEAR_ROM file on the root directory on the SD card
to clear the ROM drive was crashing. Best guess is it was trying
 to read beyond an SRAM boundary as there was a read overrun
because the write buffer was smaller than the byte count to be
written.

Creating a zeroed buffer the size as the sector size of the flash
corrects the issue.