Wide: Fix wrong DMA state after 6-byte SCSI command transfer
On ZuluSCSI Wide, after a 6-byte SCSI command the card would
sometimes crash with error
"ERROR: SCSI DMA was in state 4 when changing sync mode".

This was caused by an old bug in the synchronous wrapper around
the asynchronous read code in RP2 scsiPhy.cpp. It ended the read
by calling scsiFinishRead() with the data argument. An undocumented
assumption in the API was that final call would be with NULL argument,
like done in ZuluSCSI_disk.cpp.

In 8-bit SCSI accelerator code this made no difference, as the last
buffer completed exactly the same time as the whole transfer. Because
the 16-bit accelerator code uses second CPU core, there was a race condition
where last buffer has already finished but the whole transfer has not.