Fix >64 kB writes when using non-power-of-2 sector sizes.
By default the access wraps around the 64 kB transfer buffer
so that SD card can be written in one continuous burst. But
for non-divisible sector sizes wrapping around would split
the sector.

This commit affects only devices configured for non-power-of-2
sector size, and splits the transfer to max. 64 kB sections.